refactor: custom api function is modified
