+++ src/utility/authentication-functions.js
@@ -1,4 +1,4 @@
-import axios from "axios";
+import { apiCall } from "./api-call";
 
 const submitSignInDetails = async (
   event,
@@ -12,7 +12,7 @@ const submitSignInDetails = async (
     if (encodedToken) {
       return console.log("already logged in");
     }
-    const response = await axios.post("/api/auth/login", {
+    const response = await apiCall("post", "/api/auth/login", {
       email: email,
       password: password,
     });
@@ -39,7 +39,7 @@ const submitSignUpDetails = async (
     if (encodedToken) {
       return console.log("already logged in");
     }
-    const response = await axios.post("/api/auth/signup", {
+    const response = await apiCall("post", "/api/auth/signup", {
       firstName: firstName,
       lastName: lastName,
       email: email,
